+++ src/Model/Common/Button/Share.php
@@ -7,6 +7,9 @@ namespace Kerox\Messenger\Model\Common\Button;
 use Kerox\Messenger\Helper\UtilityTrait;
 use Kerox\Messenger\Model\Message\Attachment\Template\GenericTemplate;
 
+/**
+ * @deprecated Since version 3.2.0 and will be removed in version 4.0.0.
+ */
 class Share extends AbstractButton
 {
     use UtilityTrait;