+++ src/Model/Message/Attachment/Template/OpenGraphTemplate.php
@@ -6,6 +6,9 @@ namespace Kerox\Messenger\Model\Message\Attachment\Template;
 
 use Kerox\Messenger\Model\Message\Attachment\Template;
 
+/**
+ * @deprecated Since version 3.2.0 and will be removed in version 4.0.0.
+ */
 class OpenGraphTemplate extends Template
 {
     /**