+++ tests/TestCase/Model/Callback/MessageTest.php
@@ -11,10 +11,21 @@ class MessageTest extends AbstractTestCase
 {
     public function testMessageCallback(): void
     {
-        $message = new Message('mid.1457764197618:41d102a3e1ae206a38', 73, 'hello, world!', 'DEVELOPER_DEFINED_PAYLOAD', [['type' => 'image', 'payload' => ['url' => 'IMAGE_URL']]]);
+        $message = new Message(
+            'mid.1457764197618:41d102a3e1ae206a38',
+            'hello, world!',
+            'DEVELOPER_DEFINED_PAYLOAD',
+            [
+                [
+                    'type' => 'image',
+                    'payload' => [
+                        'url' => 'IMAGE_URL'
+                    ]
+                ]
+            ]
+        );
 
         $this->assertSame('mid.1457764197618:41d102a3e1ae206a38', $message->getMessageId());
-        $this->assertSame(73, $message->getSequence());
         $this->assertSame('hello, world!', $message->getText());
         $this->assertSame('DEVELOPER_DEFINED_PAYLOAD', $message->getQuickReply());
         $this->assertSame([['type' => 'image', 'payload' => ['url' => 'IMAGE_URL']]], $message->getAttachments());
@@ -27,7 +38,6 @@ class MessageTest extends AbstractTestCase
     {
         $message = Message::create([
             'mid' => 'mid.1457764197618:41d102a3e1ae206a38',
-            'seq' => 73,
             'text' => '',
             'quick_reply' => [
                 'payload' => '',
@@ -37,7 +47,6 @@ class MessageTest extends AbstractTestCase
         ]);
 
         $this->assertSame('mid.1457764197618:41d102a3e1ae206a38', $message->getMessageId());
-        $this->assertSame(73, $message->getSequence());
         $this->assertSame('', $message->getText());
         $this->assertSame('', $message->getQuickReply());
         $this->assertSame([], $message->getAttachments());